Move `PyFileGILRead` from `iter` to `pyfile` module
